+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "YRpp"]
 	path = YRpp
-	url = https://github.com/Phobos-developers/YRpp
+	url = https://github.com/Metadorius/YRpp
