+++ docs/Fixed-or-Improved-Logics.md
@@ -10,7 +10,6 @@ This page describes all ingame logics that are fixed or improved in Phobos witho
 - Fixed the bug that temporaryed unit cannot be erased correctly and no longer raise an error.
 - SHP debris shadows now respect the `Shadow` tag.
 - Allowed usage of TileSet of 255 and above without making NE-SW broken bridges unrepairable.
-- Adds a "Load Game" button to the retry dialog on mission failure.
 
 ![image](_static/images/turretoffset-01.png)  
 *Side offset voxel turret in Breaking Blue project*
@@ -37,7 +36,6 @@ This page describes all ingame logics that are fixed or improved in Phobos witho
   - But still some settings are ignored like `PreImpactAnim` *(Ares feature)*, this might change in future.
 - Effects like lasers are no longer drawn from wrong firing offset on weapons that use Burst.
 - Both Global Variables (`VariableNames` in `rulesmd.ini`) and Local Variables (`VariableNames` in map) are now unlimited.
-- Animations can now be offset on the X axis with `XDrawOffset`.
 
 ## Technos
