+++ docs/New-or-Enhanced-Logics.md
@@ -208,17 +208,6 @@ In `rulesmd.ini`:
 NotHuman.RandomDeathSequence=yes  ; boolean
 ```
 
-### Default disguise for individual InfantryTypes
-
-- Infantry can now have its `DefaultDisguise` overridden per-type.
-  - This tag's priority is higher than Ares' per-side `DefaultDisguise`.
-
-In `rulesmd.ini`:
-```ini
-[SOMEINFANTRY]      ; InfantryType
-DefaultDisguise=E2  ; InfantryType              
-```
-
 ## Vehicles
 
 ### Stationary vehicles
@@ -232,17 +221,6 @@ DefaultDisguise=E2  ; InfantryType
 ```ini
 [SOMETECHNO]           ; TechnoType
 NoManualMove=no        ; boolean
-```
-
-### Customizable OpenTopped Properties
-
-- You can now override settings of `OpenTopped` transport properties per TechnoType.
-
-```ini
-[SOMETECHNO]                       ; TechnoType
-OpenTopped.RangeBonus=1            ; integer
-OpenTopped.DamageMultiplier=1.3    ; float
-OpenTopped.WarpDistance=8          ; integer
 ```
 
 ## Technos
