+++ docs/Whats-New.md
@@ -45,12 +45,8 @@ New:
 - Burst delays for weapons (by Starkku)
 - PowerPlant Enhancer (by secsome)
 - Unlimited Global / Local Variables (by secsome)
-- Adds a "Load Game" button to the retry dialog on mission failure (by secsome)
-- Default disguise for individual InfantryTypes (by secsome)
 - Allow `NotHuman=yes` infantry to use random `Death` anim sequence (by Otamaa)
 - Ability for warheads to trigger specific `NotHuman=yes` infantry `Death` anim sequence (by Otamaa)
-- XDrawOffset for animations (by Morton)
-- Customizable OpenTopped properties (by Otamaa)
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)
