+++ src/Ext/BuildingType/Body.h
@@ -38,6 +38,7 @@ public:
 		}
 
 		virtual void LoadFromStream(PhobosStreamReader & Stm) override;
+
 		virtual void SaveToStream(PhobosStreamWriter & Stm) override;
 
 	private:
