+++ src/Ext/House/Body.h
@@ -5,18 +5,15 @@
 #include <Utilities/Container.h>
 #include <Utilities/TemplateDef.h>
 
-#include <Ext/BuildingType/Body.h>
-
-#include <map>
 
 class HouseExt 
 {
-public:
+public:/*
 	using base_type = HouseClass;
 	class ExtData final : public Extension<HouseClass>
 	{
 	public:
-		std::map<BuildingTypeExt::ExtData*, int> BuildingCounter;
+		
 
 		ExtData(HouseClass* OwnerObject) : Extension<HouseClass>(OwnerObject)
 		{ }
@@ -26,11 +23,8 @@ public:
 		//virtual void Initialize() override;
 		virtual void InvalidatePointer(void* ptr, bool bRemoved) override {}
 
-		virtual void LoadFromStream(PhobosStreamReader& Stm) override;
-		virtual void SaveToStream(PhobosStreamWriter& Stm) override;
-	private:
-		template <typename T>
-		void Serialize(T& Stm);
+		virtual void LoadFromStream(IStream * Stm);
+		virtual void SaveToStream(IStream * Stm);
 	};
 
 	class ExtContainer final : public Container<HouseExt> {
@@ -39,11 +33,7 @@ public:
 		~ExtContainer();
 	};
 
-	static ExtContainer ExtMap;
-
-	static bool LoadGlobals(PhobosStreamReader& Stm);
-	static bool SaveGlobals(PhobosStreamWriter& Stm);
-
+	static ExtContainer ExtMap;*/
 	static int ActiveHarvesterCount(HouseClass* pThis);
 	static int TotalHarvesterCount(HouseClass* pThis);
 	static HouseClass* GetHouseKind(OwnerHouseKind kind, bool allowRandom, HouseClass* pDefault, HouseClass* pInvoker = nullptr, HouseClass* pVictim = nullptr);
