--- src/Ext/House/Hooks.cpp
@@ -1,39 +1,0 @@
-#include "Body.h"
-
-#include <unordered_map>
-
-DEFINE_HOOK(0x508C30, HouseClass_UpdatePower_UpdateCounter, 0x5)
-{
-	GET(HouseClass*, pThis, ECX);
-	auto pHouseExt = HouseExt::ExtMap.Find(pThis);
-
-	pHouseExt->BuildingCounter.clear();
-
-	// This pre-iterating ensure our process to be done in O(NM) instead of O(N^2),
-	// as M should be much less than N, this will be a great improvement. - secsome
-	for (auto& pBld : pThis->Buildings)
-	{
-		if (pBld && !pBld->InLimbo && pBld->IsOnMap)
-		{
-			const auto pExt = BuildingTypeExt::ExtMap.Find(pBld->Type);
-			if (pExt->PowerPlantEnhancer_Buildings.size() &&
-				(pExt->PowerPlantEnhancer_Amount != 0 || pExt->PowerPlantEnhancer_Factor != 1.0f))
-			{
-				++pHouseExt->BuildingCounter[pExt];
-			}
-		}
-	}
-
-	return 0;
-}
-
-// Power Plant Enhancer #131
-DEFINE_HOOK(0x508CF2, HouseClass_UpdatePower_PowerOutput, 0x7)
-{
-	GET(HouseClass*, pThis, ESI);
-	GET(BuildingClass*, pBld, EDI);
-
-	pThis->PowerOutput += BuildingTypeExt::GetEnhancedPower(pBld, pThis);
-
-	return 0x508D07;
-}
