+++ src/Ext/Techno/Hooks.cpp
@@ -19,7 +19,7 @@ DEFINE_HOOK(0x6F9E50, TechnoClass_AI, 0x5)
 
 	// LaserTrails update routine is in TechnoClass::AI hook because TechnoClass::Draw
 	// doesn't run when the object is off-screen which leads to visual bugs - Kerbiter
-	for (auto const& trail : pExt->LaserTrails)
+	for (auto const& trail: pExt->LaserTrails)
 		trail->Update(TechnoExt::GetFLHAbsoluteCoords(pThis, trail->FLH, trail->IsOnTurret));
 
 	return 0;
@@ -126,7 +126,7 @@ DEFINE_HOOK(0x518505, InfantryClass_TakeDamage_NotHuman, 0x4)
 	REF_STACK(args_ReceiveDamage const, receiveDamageArgs, STACK_OFFS(0xD0, -0x4));
 
 	// Die1-Die5 sequences are offset by 10
-	constexpr auto Die = [](int x) { return x + 10; };
+	#define Die(x) x + 10
 
 	int resultSequence = Die(1);
 	auto const pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
@@ -144,71 +144,10 @@ DEFINE_HOOK(0x518505, InfantryClass_TakeDamage_NotHuman, 0x4)
 		}
 	}
 
+	#undef Die(x)
+
 	R->ECX(pThis);
 	pThis->PlayAnim(static_cast<Sequence>(resultSequence), true);
 
 	return 0x518515;
-}
-
-// Customizable OpenTopped Properties
-// Author: Otamaa
-
-DEFINE_HOOK(0x6F72D2, TechnoClass_IsCloseEnoughToTarget_OpenTopped_RangeBonus, 0xC)
-{
-	GET(TechnoClass* const, pThis, ESI);
-
-	if (auto pTransport = pThis->Transporter)
-	{
-		if (auto pExt = TechnoTypeExt::ExtMap.Find(pTransport->GetTechnoType()))
-		{
-			R->EAX(pExt->OpenTopped_RangeBonus.Get(RulesClass::Instance->OpenToppedRangeBonus));
-			return 0x6F72DE;
-		}
-	}
-
-	return 0;
-}
-
-DEFINE_HOOK(0x6FE43B, TechnoClass_Fire_OpenTopped_DmgMult, 0x8)
-{
-	enum { ApplyDamageMult = 0x6FE45A, ContinueCheck = 0x6FE460 };
-	
-	GET(TechnoClass* const, pThis, ESI);
-
-	//replacing whole check due to `fild`
-	if (pThis->InOpenToppedTransport)
-	{
-		GET_STACK(int, nDamage, STACK_OFFS(0xB4, -0x2C));
-		float nDamageMult = static_cast<float>(RulesClass::Instance->OpenToppedDamageMultiplier);
-
-		if (auto pTransport = pThis->Transporter)
-		{
-			if (auto pExt = TechnoTypeExt::ExtMap.Find(pTransport->GetTechnoType()))
-			{
-				//it is float isnt it YRPP ? , check tomson26 YR-IDB !
-				nDamageMult = pExt->OpenTopped_DamageMultiplier.Get(nDamageMult);
-			}
-		}
-
-		R->EAX(Game::F2I(nDamage * nDamageMult));
-		return ApplyDamageMult;
-	}
-
-	return ContinueCheck;
-}
-
-DEFINE_HOOK(0x71A82C, TemporalClass_AI_Opentopped_WarpDistance, 0xC)
-{
-	GET(TemporalClass* const, pThis, ESI);
-
-	if (auto pTransport = pThis->Owner->Transporter)
-	{
-		if (auto pExt = TechnoTypeExt::ExtMap.Find(pTransport->GetTechnoType()))
-		{
-			R->EDX(pExt->OpenTopped_WarpDistance.Get(RulesClass::Instance->OpenToppedWarpDistance));
-			return 0x71A838;
-		}
-	}
-
-	return 0;
 }
