+++ src/Ext/TechnoType/Body.h
@@ -61,12 +61,6 @@ public:
 		Valueable<bool> DestroyAnim_Random;
 		Valueable<bool> NotHuman_RandomDeathSequence;
 
-		Nullable<InfantryTypeClass*> DefaultDisguise;
-
-		Nullable<int> OpenTopped_RangeBonus;
-		Nullable<float> OpenTopped_DamageMultiplier;
-		Nullable<int> OpenTopped_WarpDistance;
-
 		struct LaserTrailDataEntry
 		{
 			ValueableIdx<LaserTrailTypeClass> idxType;
@@ -119,11 +113,7 @@ public:
 			OreGathering_FramesPerDir(),
 			LaserTrailData(),
 			DestroyAnim_Random(true),
-			NotHuman_RandomDeathSequence(false),
-			DefaultDisguise(),
-			OpenTopped_RangeBonus(),
-			OpenTopped_DamageMultiplier(),
-			OpenTopped_WarpDistance()
+			NotHuman_RandomDeathSequence(false)
 		{ }
 
 		virtual ~ExtData() = default;
