+++ src/Ext/TechnoType/Hooks.cpp
@@ -1,5 +1,4 @@
 #include <UnitClass.h>
-#include <InfantryClass.h>
 #include <BuildingClass.h>
 #include <ScenarioClass.h>
 #include <HouseClass.h>
@@ -156,28 +155,6 @@ DEFINE_HOOK(0x700C58, TechnoClass_CanPlayerMove_NoManualMove, 0x6)
 
 	if (auto pExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
 		return pExt->NoManualMove ? 0x700C62 : 0;
-
-	return 0;
-}
-
-DEFINE_HOOK_AGAIN(0x522790, TechnoClass_DefaultDisguise, 0x6) // InfantryClass_SetDisguise_DefaultDisguise
-DEFINE_HOOK(0x6F421C, TechnoClass_DefaultDisguise, 0x6) // TechnoClass_DefaultDisguise
-{
-	GET(TechnoClass*, pThis, ESI);
-
-	enum { SetDisguise = 0x5227BF, DefaultDisguise = 0x6F4277 };
-
-	if (auto const pExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType()))
-	{
-		if (pExt->DefaultDisguise.isset())
-		{
-			pThis->Disguise = pExt->DefaultDisguise;
-			pThis->Disguised = true;
-			return R->Origin() == 0x522790 ? SetDisguise : DefaultDisguise;
-		}
-	}
-
-	pThis->Disguised = false;
 
 	return 0;
 }
